add solution building

+++ src/entities/vector.h
@@ -8,14 +8,14 @@ class Vector {
     std::vector<int> canonized;
 public:
     Vector(const std::vector<int>& values);
-    Vector(int dimension, int index);
+    Vector(size_t dimension, size_t index);
 
     Vector operator+(const Vector& vector) const;
     Vector operator-(const Vector& vector) const;
     Vector operator-() const;
     Vector operator*(int scale) const;
 
-    int operator[](int index) const;
+    int operator[](size_t index) const;
     size_t getHash() const;
 
     bool operator==(const Vector& vector) const;
@@ -23,11 +23,11 @@ public:
     bool operator<(const Vector& vector) const;
     int compare(const Vector& vector) const; 
 
-    int getDimension() const;
-    int getSupport() const;
+    size_t getDimension() const;
+    size_t getSupport() const;
     int getMaxAbs() const;
-    int getHammingDistance(const Vector& vector) const;
-    int getMatchesCount(const Vector& vector) const;
+    size_t getHammingDistance(const Vector& vector) const;
+    size_t getMatchesCount(const Vector& vector) const;
 private:
     void canonize();
 };
